+++ tests/wd/configs/testing.py
@@ -3,3 +3,4 @@ from configs.default import *
 
 TESTING = True
 DEBUG = True
+PORT = 4000